Register and Login Functioning
Just need to make it so it shows an incorrect password onscreen for the
user, as well as whether or not their email has been used already. Right
now the backend is sneding a resonse to the front end that will log out
either 'correct' or incorrect' for the password.